# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PeacePay API
  description: |
    PeacePay - منصة الدفع الآمن للتجارة الإلكترونية في مصر
    
    Secure Escrow Payment Platform for E-commerce in Egypt
    
    ## Authentication
    All protected endpoints require a Bearer token obtained from the login endpoint.
    ```
    Authorization: Bearer {token}
    ```
    
    ## Response Format
    All responses follow a consistent format:
    ```json
    {
      "success": true|false,
      "message": "رسالة توضيحية",
      "data": { ... },
      "errors": { ... }  // Only on validation errors
    }
    ```
    
    ## Error Codes
    - `400` - Bad Request
    - `401` - Unauthorized
    - `403` - Forbidden
    - `404` - Not Found
    - `422` - Validation Error
    - `429` - Too Many Requests
    - `500` - Server Error
    
  version: 1.0.0
  contact:
    name: PeacePay Support
    email: support@peacepay.eg
  license:
    name: Proprietary
servers:
  - url: https://api.peacepay.eg/api/v1
    description: Production Server
  - url: https://staging-api.peacepay.eg/api/v1
    description: Staging Server
  - url: http://localhost:8000/api/v1
    description: Local Development
...
paths:
...
  /peacelinks/{id}:
    get:
      tags: [PeaceLink]
      summary: Get PeaceLink details
      operationId: getPeaceLink
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: UUID or reference
      responses:
        '200':
          description: PeaceLink details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeaceLinkDetails'
        '404':
          description: PeaceLink not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PeaceLink:
      type: object
      properties:
        id:
          type: string
        reference:
          type: string
        status:
          type: string
          enum: [pending, funded, dsp_assigned, in_transit, delivered, released, cancelled, disputed, refunded]
        product_description:
          type: string
        item_amount:
          type: number
        platform_fee:
          type: number
        delivery_fee:
          type: number
        total_amount:
          type: number
        buyer:
          $ref: '#/components/schemas/UserPublic'
        merchant:
          $ref: '#/components/schemas/UserPublic'
        created_at:
          type: string
          format: date-time
    PeaceLinkDetails:
      allOf:
        - $ref: '#/components/schemas/PeaceLink'
        - type: object
          properties:
            dsp:
              $ref: '#/components/schemas/UserPublic'
            delivery_address:
              type: string
            delivery_city:
              type: string
            tracking_code:
              type: string
            timeline:
              type: array
              items:
                type: object
                properties:
                  event:
                    type: string
                  timestamp:
                    type: string
                  actor:
                    type: string
            can_cancel:
              type: boolean
            can_confirm_delivery:
              type: boolean
            can_dispute:
              type: boolean
    UserPublic:
      type: object
      properties:
        name:
          type: string
        masked_phone:
          type: string
...